Archiving meetings in our community page
Instead of using google docs, we discussed in the last meeting
to add that into a markdown format in our docs.

This does not fix an actual problem, more a convenience way
to document our meetings with a code review approach.

This fixes it by adding the meetings into the community page,
rather than using a different page. This content was copied
verbatim from the google docs.

Signed-off-by: Jean-Philippe Evrard <open-source@a.spamming.party>