Add how to prononce kured
While it does not bother Daniel much, I have been asking for
this for years, as it's been bothering me.

We finally decided in the last meeting for a way to say "kured".

This fixes it by documenting in the FAQ, as it is probably not
worth putting on the front page anyway (we did not care for it
in the last 4 years, it should not start now...)

Signed-off-by: Jean-Philippe Evrard <open-source@a.spamming.party>